pyproject.toml
[project]
name = "ultimate-ai-assistant"
version = "1.0.0"
description = "ZORA - Multi-modal AI assistant with voice interaction, persistent memory, and real-world integrations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.5",                              # Agent-to-Agent protocol support
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "authlib==1.5.2",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "cryptography==44.0.3",
    "deprecated==1.2.18",
    "docstring-parser==0.16",
    "fastapi==0.115.12",
    "google-adk==1.8.0",
    "google-api-core==2.24.2",
    "google-api-python-client==2.169.0",
    "google-auth==2.40.1",
    "google-crc32c==1.7.1",
    "google-genai==1.21.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.20.3",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "mcp==1.8.1",
    "nest-asyncio>=1.6.0",
    "numpy==2.2.5",
    "oauthlib==3.2.2",
    "opentelemetry-api==1.33.0",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.33.0",
    "opentelemetry-semantic-conventions==0.54b0",
    "packaging==25.0",
    "pdfplumber>=0.10.0",
    "pillow>=10.0.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pymupdf>=1.23.0",
    "pyparsing==3.2.3",
    "pytesseract>=0.3.10",
    "python-dateutil==2.9.0.post0",
    "python-docx>=0.8.11",
    "python-dotenv==1.1.0",
    "python-multipart>=0.0.6",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "shapely==2.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "sse-starlette==2.3.4",
    "starlette==0.46.2",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "tzlocal==5.3.1",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "zep-cloud==2.12.1",
    "zep-python==2.0.2",
    "zipp==3.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "httpx>=0.24.0",          # For testing HTTP endpoints
    "pytest-mock>=3.10.0",    # For mocking in tests
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "zep_cloud.*",
    "zep_python.*",
    "pytesseract.*",
    "pdfplumber.*",
    "pymupdf.*",
]
ignore_missing_imports = true
